streamlit/requirements.txt
streamlit
mysql-connector-python
bcrypt
pandas
plotly
